FunctionApps/YamlSyntaxValidator/requirements.txt
azure-functions
pyyaml